+++ tests/test_exceptions.py
@@ -13,7 +13,6 @@ _SUBCLASSES = [
     "VeoliaAPIResponseError",
     "VeoliaAPIGetDataError",
     "VeoliaAPISetDataError",
-    "VeoliaAPIUnknownError",
     "VeoliaAPIRateLimitError",
 ]